requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-filter>=23.2
django-cors-headers>=4.2.0
Pillow>=10.0.0
openpyxl>=3.1.2
python-docx>=0.8.11
reportlab>=4.0.4
pandas>=2.0.3
numpy>=1.24.3
celery>=5.3.1
redis>=4.6.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
